Fetch recipients from Forms API

+++ packages/bygger/src/hooks/useFormsApiRecipients.ts
@@ -1,5 +1,5 @@
-import { http as baseHttp, useAppConfig } from '@navikt/skjemadigitalisering-shared-components';
-import { Recipient } from '@navikt/skjemadigitalisering-shared-domain';
+import { http as baseHttp, NavFormioJs, useAppConfig } from '@navikt/skjemadigitalisering-shared-components';
+import { NavFormType, Recipient } from '@navikt/skjemadigitalisering-shared-domain';
 import { useFeedbackEmit } from '../context/notifications/FeedbackContext';
 
 const useFormsApiRecipients = () => {
@@ -46,11 +46,28 @@ const useFormsApiRecipients = () => {
 
   const deleteRecipient = async (recipientId: string): Promise<void> => {
     try {
+      const forms = await getFormsWithMottaksadresse(recipientId);
+      if (forms.length > 0) {
+        const skjemanummerliste = forms.map((form) => form.properties.skjemanummer).join(', ');
+        feedbackEmit.error(`Mottakeren brukes i følgende skjema: ${skjemanummerliste}`);
+        return;
+      }
       return await http.delete(`${baseUrl}/${recipientId}`);
     } catch (error) {
       const message = (error as Error)?.message;
       feedbackEmit.error(`Feil ved sletting av mottaker. ${message}`);
     }
+  };
+
+  // Verify that no form refers to the recipient before deleting it. This check should be done in the Forms API when
+  // form definitions are moved there too.
+  const getFormsWithMottaksadresse = async (recipientId: string): Promise<NavFormType[]> => {
+    return fetch(
+      `${NavFormioJs.Formio.getProjectUrl()}/form?type=form&tags=nav-skjema&limit=1000&properties.mottaksadresseId=${recipientId}`,
+      {
+        method: 'GET',
+      },
+    ).then((forms) => forms.json());
   };
 
   return {
